+++ api/src/team/application/routes.js
@@ -1,5 +1,6 @@
 import { certificationCenterInvitationAdminRoutes } from './certification-center-invitation/certification-center-invitation.admin.route.js';
 import { certificationCenterInvitationRoutes } from './certification-center-invitation/certification-center-invitation.route.js';
+import { organizationInvitationRoutes } from './organization-invitations/organization-invitation.route.js';
 import { prescriberInformationsRoute } from './prescriber-informations.route.js';
 
 const register = async function (server) {
@@ -7,6 +8,7 @@ const register = async function (server) {
     ...certificationCenterInvitationRoutes,
     ...certificationCenterInvitationAdminRoutes,
     ...prescriberInformationsRoute,
+    ...organizationInvitationRoutes,
   ]);
 };
